:truck: move `SideBar` to `Panel/`
