--- app/components/Bar/index.js
@@ -1,1 +1,0 @@
-export const SideBar = require('./SideBar').default
