+++ app/components/Panel/SideBar.jsx
